Migrate focussed item to ts
